outsource to DragToScrollGesture.swift and clean up
